refactor: move broker to shared pkg
