+++ go.mod
@@ -4,6 +4,7 @@ go 1.22.4
 
 require (
 	github.com/joho/godotenv v1.5.1
+	github.com/rabbitmq/amqp091-go v1.10.0
 	github.com/robfig/cron/v3 v3.0.1
 	github.com/stretchr/testify v1.9.0
 	gorm.io/driver/postgres v1.5.9
